requirements.txt
# This is a generated file; DO NOT EDIT!
#
# Please edit 'setup.cfg' to add top-level dependencies and use
# './contrib/refresh-requirements.sh to regenerate this file

amqp==2.6.1
asgiref==3.3.1
attrs==20.3.0
billiard==3.6.3.0
cachetools==4.2.0
celery==4.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.5
django-allauth==0.44.0
django-appconf==1.0.4
django-avatar==5.0.0
django-bootstrap-modal-forms==2.0.1
django-compressor==2.4
django-countries==6.1.3
django-crispy-forms==1.10.0
django-csp==3.7
django-debug-toolbar==3.2
django-filter==2.4.0
django-fontawesome-5==1.0.18
django-jsonfield==1.4.1
django-redis-cache==3.0.0
django-shortuuidfield==0.1.3
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
dnspython==1.16.0
drf-spectacular==0.12.0
ecdsa==0.14.1
enum34==1.1.10
eventlet==0.29.1
frozendict==1.2
greenlet==0.4.17
gunicorn==19.9.0
h5py==3.1.0
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
influxdb==5.3.1
jsonschema==3.2.0
kaitaistruct==0.9
kombu==4.6.11
Logbook==1.5.3
lxml==4.6.2
Markdown==3.3.3
msgpack==1.0.2
mysqlclient==2.0.3
numpy==1.19.5
oauthlib==3.1.0
Pillow==8.1.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.0.0
PyLD==2.0.3
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.15.0
python-jose==3.2.0
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
rcssmin==1.0.6
redis==3.5.3
Represent==1.6.0.post0
requests==2.24.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
rsa==4.6
rush==2018.12.1
satellitetle==0.10.1
satnogs-decoders~=1.0
sentry-sdk==0.19.5
sgp4==2.14
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
spacetrack==0.16.0
sqlparse==0.4.1
Unipath==1.1
uritemplate==3.0.1
urllib3==1.25.11
vine==1.3.0
zipp==3.4.0
